Refactor Settings/Backup container. Move proptypes to top of settings component.
