--- src/components/Settings/Backup/Backup.js
@@ -1,83 +1,0 @@
-import React, { PureComponent } from 'react';
-import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
-import Button from 'material-ui/Button';
-import List, {
-  ListItem,
-  ListItemText,
-  ListItemSecondaryAction
-} from 'material-ui/List';
-
-import FullScreenDialog from '../../FullScreenDialog';
-import messages from './Backup.messages';
-
-export class Backup extends PureComponent {
-  static propTypes = {
-    boards: PropTypes.array.isRequired,
-    open: PropTypes.bool,
-    onRequestClose: PropTypes.func
-  };
-
-  handleExportClick = () => {
-    const exportFilename = 'board.json';
-    const { boards } = this.props;
-    const jsonData = new Blob([JSON.stringify(boards)], {
-      type: 'text/json;charset=utf-8;'
-    });
-
-    // IE11 & Edge
-    if (navigator.msSaveBlob) {
-      navigator.msSaveBlob(jsonData, exportFilename);
-    } else {
-      // In FF link must be added to DOM to be clicked
-      const link = document.createElement('a');
-      link.href = window.URL.createObjectURL(jsonData);
-      link.setAttribute('download', exportFilename);
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-    }
-  };
-
-  render() {
-    const { open, onRequestClose, onImport } = this.props;
-
-    return (
-      <div className="Backup">
-        <FullScreenDialog
-          open={open}
-          title={<FormattedMessage {...messages.backup} />}
-          onRequestClose={onRequestClose}
-        >
-          <List>
-            <ListItem>
-              <ListItemText
-                primary={<FormattedMessage {...messages.backup} />}
-                secondary="Backup your boards"
-              />
-              <ListItemSecondaryAction>
-                <Button onClick={this.handleExportClick}>
-                  <FormattedMessage {...messages.export} />
-                </Button>
-                <Button component="span">
-                  <label htmlFor="file">
-                    <FormattedMessage {...messages.restore} />
-                  </label>
-                  <input
-                    accept=".json,text/json,application/json"
-                    id="file"
-                    type="file"
-                    style={{ display: 'none' }}
-                    onChange={e => onImport(e)}
-                  />
-                </Button>
-              </ListItemSecondaryAction>
-            </ListItem>
-          </List>
-        </FullScreenDialog>
-      </div>
-    );
-  }
-}
-
-export default Backup;
