+++ src/components/Settings/Backup/index.js
@@ -1,1 +1,1 @@
-export { default } from './Backup';
+export { default } from './Backup.container';
